# movies_api_spec.yaml
openapi: 3.0.2
info:
  version: 2.0.0
  title: API de gestion de collection de films
  description: |
    API REST pour gérer une collection personnelle de films avec réalisateurs, genres et collections personnalisées.
    Cette API suit les principes HATEOAS (Hypermedia As The Engine Of Application State).
  contact:
    name: "Allegre Bastien"
    email: bastien.allegre@etu.univ-grenoble-alpes.fr
servers:
- url: http://localhost:3000/api
  description: Serveur de développement local
paths:
...
  /genres:
    get:
      tags:
        - Genres
      summary: Liste tous les genres
      description: Récupère la liste complète des genres
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Liste des genres
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  genres:
                    type: array
                    items:
                      $ref: '#/components/schemas/GenreWithLinks'
                  _links:
                    $ref: '#/components/schemas/PaginationLinks'
    post:
      tags:
        - Genres
      summary: Crée un nouveau genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreInput'
      responses:
        "201":
          description: Genre créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreWithLinks'
  /genres/{id}:
    get:
      tags:
        - Genres
      summary: Récupère un genre
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "200":
          description: Genre retourné
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreDetailWithLinks'
        "404":
          description: Genre non trouvé
    put:
      tags:
        - Genres
      summary: Met à jour un genre
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreInput'
      responses:
        "200":
          description: Genre mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreWithLinks'
    delete:
      tags:
        - Genres
      summary: Supprime un genre
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "204":
          description: Genre supprimé
        "404":
          description: Genre non trouvé
  /genres/{id}/movies:
    get:
      tags:
        - Genres
      summary: Liste les films d'un genre
      description: Récupère tous les films appartenant à un genre
      parameters:
        - $ref: '#/components/parameters/idParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Liste des films du genre
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    type: object
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'
                      genre:
                        $ref: '#/components/schemas/Link'

  # ==================== DIRECTORS ====================
  /directors:
    get:
      tags:
        - Directors
      summary: Liste tous les réalisateurs
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: sort
          in: query
          description: Tri par nom ou nombre de films
          schema:
            type: string
            enum: [nom, nombre_films]
      responses:
        "200":
          description: Liste des réalisateurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  directors:
                    type: array
                    items:
                      $ref: '#/components/schemas/DirectorWithLinks'
                  _links:
                    $ref: '#/components/schemas/PaginationLinks'
    post:
      tags:
        - Directors
      summary: Ajoute un nouveau réalisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorInput'
      responses:
        "201":
          description: Réalisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorWithLinks'
  /directors/{id}:
    get:
      tags:
        - Directors
      summary: Récupère un réalisateur
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "200":
          description: Réalisateur retourné avec sa filmographie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorDetailWithLinks'
        "404":
          description: Réalisateur non trouvé
    put:
      tags:
        - Directors
      summary: Met à jour un réalisateur
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorInput'
      responses:
        "200":
          description: Réalisateur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorWithLinks'
    delete:
      tags:
        - Directors
      summary: Supprime un réalisateur
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "204":
          description: Réalisateur supprimé
        "404":
          description: Réalisateur non trouvé
  /directors/{id}/movies:
    get:
      tags:
        - Directors
      summary: Liste les films d'un réalisateur
      description: Récupère la filmographie complète d'un réalisateur
      parameters:
        - $ref: '#/components/parameters/idParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Filmographie du réalisateur
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    type: object
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'
                      director:
                        $ref: '#/components/schemas/Link'

  # ==================== STATS ====================
...
components:
  parameters:
    idParam:
      name: id
      in: path
      required: true
      description: Identifiant unique de la ressource
      schema:
        type: string
    pageParam:
      name: page
      in: query
      description: Numéro de page (commence à 0)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    limitParam:
      name: limit
      in: query
      description: Nombre d'éléments par page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    MovieInput:
      type: object
      required:
        - titre
        - annee
      properties:
        titre:
          type: string
          example: "Inception"
        annee:
          type: integer
          minimum: 1888
          example: 2010
        director_id:
          type: string
          description: ID du réalisateur
        genre_ids:
          type: array
          items:
            type: string
          description: Liste des IDs de genres
        duree:
          type: integer
          description: Durée en minutes
          example: 148
        synopsis:
          type: string
        statut:
          type: string
          enum: [a_voir, vu, en_cours]
          default: "a_voir"
        note:
          type: number
          format: float
          minimum: 0
          maximum: 10
        commentaire:
          type: string
        affiche_url:
          type: string
          format: uri
        tmdb_id:
          type: integer
        tags:
          type: array
          items:
            type: string
    Movie:
      allOf:
        - $ref: '#/components/schemas/MovieInput'
        - type: object
          required:
            - id
            - date_ajout
          properties:
            id:
              type: string
            date_ajout:
              type: string
              format: date-time
            date_visionnage:
              type: string
              format: date-time
    MovieWithLinks:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                director:
                  $ref: '#/components/schemas/Link'
                genres:
                  type: array
                  items:
                    $ref: '#/components/schemas/Link'
                collections:
                  type: array
                  items:
                    $ref: '#/components/schemas/Link'
...
    GenreInput:
      type: object
      required:
        - nom
      properties:
        nom:
          type: string
          example: "Science-Fiction"
        description:
          type: string
    Genre:
      allOf:
        - $ref: '#/components/schemas/GenreInput'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
            nombre_films:
              type: integer
    GenreWithLinks:
      allOf:
        - $ref: '#/components/schemas/Genre'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
    GenreDetailWithLinks:
      allOf:
        - $ref: '#/components/schemas/Genre'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
                update:
                  $ref: '#/components/schemas/Link'
                delete:
                  $ref: '#/components/schemas/Link'

    # ==================== DIRECTOR SCHEMAS ====================
    DirectorInput:
      type: object
      required:
        - nom
        - prenom
      properties:
        nom:
          type: string
          example: "Nolan"
        prenom:
          type: string
          example: "Christopher"
        date_naissance:
          type: string
          format: date
        nationalite:
          type: string
        biographie:
          type: string
        photo_url:
          type: string
          format: uri
    Director:
      allOf:
        - $ref: '#/components/schemas/DirectorInput'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
            nombre_films:
              type: integer
    DirectorWithLinks:
      allOf:
        - $ref: '#/components/schemas/Director'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
    DirectorDetailWithLinks:
      allOf:
        - $ref: '#/components/schemas/Director'
        - type: object
          properties:
            filmographie:
              type: array
              items:
                $ref: '#/components/schemas/MovieWithLinks'
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
                update:
                  $ref: '#/components/schemas/Link'
                delete:
                  $ref: '#/components/schemas/Link'

    # ==================== STATS SCHEMAS ====================
...
    Link:
      type: object
      required:
        - href
      properties:
        href:
          type: string
          format: uri
          example: "/api/movies/123"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          default: "GET"
        rel:
          type: string
          example: "self"
    PaginationLinks:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'
        first:
          $ref: '#/components/schemas/Link'
        prev:
          $ref: '#/components/schemas/Link'
        next:
          $ref: '#/components/schemas/Link'
        last:
          $ref: '#/components/schemas/Link'
...